+++ data/options_core.lua
@@ -42,9 +42,9 @@ function SQP:CreateOptionsPanel()
     
     -- Create preview section at the top
     local previewContainer = CreateFrame("Frame", nil, container, "BackdropTemplate")
-    previewContainer:SetHeight(180)
-    previewContainer:SetPoint("TOPLEFT", header, "BOTTOMLEFT", 0, -10)
-    previewContainer:SetPoint("TOPRIGHT", header, "BOTTOMRIGHT", 0, -10)
+    previewContainer:SetHeight(148)
+    previewContainer:SetPoint("TOPLEFT", header, "BOTTOMLEFT", 0, -6)
+    previewContainer:SetPoint("TOPRIGHT", header, "BOTTOMRIGHT", 0, -6)
     previewContainer:SetBackdrop(self.BACKDROP_DARK)
     previewContainer:SetBackdropColor(0.08, 0.08, 0.08, 0.8)
     previewContainer:SetBackdropBorderColor(unpack(self.SECTION_COLOR))